pyproject.toml
[tool.poetry]
name = "Up-Down-Gambling"
version = "0.1.0"
description = ""
authors = ["AncientBison"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.13.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
